+++ lmcache/v1/multiprocess/modules/lmcache_driven_transfer.py
@@ -3,56 +3,402 @@
 
 # Standard
 from dataclasses import dataclass
-import threading
+from itertools import islice
+from typing import Generator, Sequence
 import time
 
 # Third Party
 import torch
 
 # First Party
+from lmcache import torch_dev, torch_device_type
 from lmcache.logging import init_logger
-from lmcache.utils import _lmcache_nvtx_annotate
+from lmcache.utils import (
+    EngineType,
+    _lmcache_nvtx_annotate,
+    check_interprocess_event_support,
+)
 from lmcache.v1.distributed.api import (
     MemoryLayoutDesc,
     ObjectKey,
 )
+from lmcache.v1.gpu_connector.gpu_ops import (
+    lmcache_memcpy_async_d2h,
+    lmcache_memcpy_async_h2d,
+)
+from lmcache.v1.gpu_connector.utils import LayoutHints
+from lmcache.v1.memory_management import MemoryObj
+from lmcache.v1.mp_observability.event import Event, EventType
 from lmcache.v1.multiprocess.custom_types import (
     IPCCacheServerKey,
-    RegisterNonGpuContextPayload,
+    KVCache,
 )
-from lmcache.v1.multiprocess.engine_context import MPCacheServerContext, ShmPoolInfo
+from lmcache.v1.multiprocess.engine_context import MPCacheServerContext
 from lmcache.v1.multiprocess.engine_module import (
     HandlerSpec,
     ThreadPoolType,
 )
+from lmcache.v1.multiprocess.gpu_context import GPUCacheContext
+from lmcache.v1.multiprocess.group_view import EngineGroupInfo
+from lmcache.v1.multiprocess.native_completion import (
+    DeviceHostFuncDispatcher,
+    submit_callback_to_stream,
+)
 from lmcache.v1.multiprocess.protocols.base import RequestType
-from lmcache.v1.multiprocess.protocols.engine import (
-    PrepareRetrieveResponse,
-    PrepareStoreResponse,
-    RegisterNonGpuContextResponse,
-)
-from lmcache.v1.multiprocess.transfer_context.base import NonGpuContextMetadata
-
-# Local
-from .server_transfer import (
-    TransferStrategy,
-    create_transfer_strategy,
-)
+from lmcache.v1.platform.cache_context import create_cache_context
+import lmcache.c_ops as lmc_ops
 
 logger = init_logger(__name__)
 
 
-@dataclass
-class NonGPUContextEntry:
-    """Registered non-GPU context metadata for a single worker instance.
+def get_layout_desc(
+    gpu_context: GPUCacheContext,
+    num_tokens: int,
+    object_group_id: int,
+) -> MemoryLayoutDesc:
+    """Get the memory layout description for a specific object group.
 
-    Attributes:
-        metadata: Layout metadata describing the non-CUDA chunk format.
-        model_name: The name of the model associated with this context.
-        world_size: The world size associated with this context.
+    The returned layout describes the single memory object that backs
+    ``object_group_id``: one (shape, dtype) entry per kernel group in that
+    object group, in the kernel groups' declared layout order. Kernel groups
+    may have different shapes and dtypes.
+
+    Args:
+        cache_context: The GPU cache context containing the KV cache information.
+        num_tokens: The number of tokens to determine the layout for.
+        object_group_id: Index of the object group whose layout to build.
+
+    Returns:
+        MemoryLayoutDesc: The memory layout description containing shapes and
+        dtypes, one entry per kernel group in the object group.
+    """
+    object_group = gpu_context.kv_layer_groups_manager.object_groups[object_group_id]
+    shapes_and_dtypes = [
+        gpu_context.get_kernel_group_shape_dtype(num_tokens, kernel_group_idx)
+        for kernel_group_idx in object_group.kernel_group_indices
+    ]
+    shapes, dtypes = zip(*shapes_and_dtypes, strict=False)
+    return MemoryLayoutDesc(shapes=list(shapes), dtypes=list(dtypes))
+
+
+def batched_iteration_with_skip(
+    lst: Sequence,
+    batch_size: int,
+    skip_count: int,
+) -> Generator[tuple[int, tuple], None, None]:
+    """Utility function to iterate over a list in batches with an initial skip.
+
+    Args:
+        lst: The list to iterate over.
+        batch_size: The size of each batch.
+        skip_count: The number of items to skip at the start of the list.
+
+    Yields:
+        Tuples of (batch_start_idx, batch) where batch is a tuple of items
+        from the list, and batch_start_idx is the "original" index of the first
+        item in the batch.
+
+    Raises:
+        ValueError: If batch_size is less than 1 or skip_count is negative.
+
+    Note:
+        Batch_idx is the index of the batch in the original list, accounting
+        for the skipped items. For example, if skip_count is 10 and batch_size
+        is 5, the first yielded batch will have batch_start_idx=10.
+    """
+    if batch_size < 1:
+        raise ValueError("batch size must be at least one")
+    if skip_count < 0:
+        raise ValueError("skip_count must be non-negative")
+
+    it = iter(lst)
+    # Skip the initial items
+    for _ in range(skip_count):
+        next(it, None)
+    batch_start_idx = skip_count
+    while batch := tuple(islice(it, batch_size)):
+        yield batch_start_idx, batch
+        batch_start_idx += len(batch)
+
+
+def downsample_and_stage_block_ids(
+    cache_context: GPUCacheContext,
+    block_ids: list[list[int]],
+) -> list[torch.Tensor]:
+    """Cut the block id lists to skip the unneeded blocks in a chunk and
+    stage it into GPU tensors for later use.
+
+    This mainly targets the case where a portion of the blocks are not
+    needed for every chunk, such as deepseek v4's swa cache.
+
+    Note that the we do NOT do any object-level skipping here.
+
+    Args:
+        cache_context: The GPU cache context containing the KV cache information.
+        block_ids: The original block id lists, indexed by LMCache KV group index.
+
+    Returns:
+        The cut block id lists, indexed by LMCache KV group index.
+
+    Note:
+        This function has some coupled logic with transfer_kv_per_object_group below.
+        The caller need to make sure that the block ids seen by
+        transfer_kv_per_object_group are produced by this function.
+
+    Example:
+        If a model have 2 kernel groups, one is full attention with block size 32,
+        one is swa attention with block size 32 and sliding window size 64, and
+        LMCache has a chunk size of 128. And there are 2 chunks in total (256 tokens).
+
+        The input will be:
+        [
+          [1, 2, 3, 4, 5, 6, 7, 8],  # block ids for the full attention group
+          [11, 12, 13, 14, 15, 16, 17, 18], # block ids for the swa attention group
+        ]
+
+        The output will be
+        [
+          [1, 2, 3, 4, 5, 6, 7, 8],  # full attention group still needs all block ids
+          [13, 14, 17, 18], # swa attention group only needs the last 2 block per chunk
+        ]
+    """
+    num_kernel_groups = cache_context.kv_layer_groups_manager.num_kernel_groups
+    for kernel_group_id in range(num_kernel_groups):
+        subchunk_sw_size_tokens = (
+            cache_context.kv_layer_groups_manager.get_subchunk_sw_size_tokens(
+                kernel_group_id
+            )
+        )
+        tokens_per_chunk = min(
+            cache_context.lmcache_tokens_per_chunk, subchunk_sw_size_tokens
+        )
+        keep_blocks_per_chunk = cache_context.calculate_num_blocks(
+            tokens_per_chunk, kernel_group_id
+        )
+        total_blocks_per_chunk = cache_context.calculate_num_blocks(
+            cache_context.lmcache_tokens_per_chunk, kernel_group_id
+        )
+
+        new_block_ids = []
+        old_block_ids = block_ids[kernel_group_id]
+        assert len(old_block_ids) % total_blocks_per_chunk == 0, (
+            f"len(block_ids[{kernel_group_id}]) should be a multiple "
+            f"of total_blocks_per_chunk ({total_blocks_per_chunk}), but got "
+            f"{len(old_block_ids)}"
+        )
+
+        for i in range(0, len(old_block_ids), total_blocks_per_chunk):
+            chunk_block_ids = old_block_ids[i : i + total_blocks_per_chunk]
+            new_block_ids.extend(chunk_block_ids[-keep_blocks_per_chunk:])
+
+        block_ids[kernel_group_id] = new_block_ids
+
+    # Stage the cut block ids into GPU tensors
+    block_ids_gpu = cache_context.copy_view_block_ids_to_gpu(block_ids)
+    return block_ids_gpu
+
+
+def _recalculate_blocks_to_skip(
+    blocks_per_chunk: int,
+    blocks_per_window: int,
+    blocks_to_skip: int,
+) -> int:
+    """Re-calculate the number of blocks to skip for a batch of chunks based
+    on the blocks per chunk and blocks per sliding window WHEN the window
+    size is smaller than the lmcache chunk size.
+
+    Args:
+        blocks_per_chunk: The total number of blocks in one chunk for the
+            current group.
+        blocks_per_window: The number of blocks in the sliding window
+            for the current group. Should be less than or equal to
+            blocks_per_chunk.
+        blocks_to_skip: The number of blocks to skip.
+
+    Returns:
+        The re-calculated number of blocks to skip for the current batch of
+        chunks.
+    """
+    if blocks_per_chunk == blocks_per_window:
+        return blocks_to_skip
+
+    full_windows_to_skip = blocks_to_skip // blocks_per_chunk
+    tail_blocks = blocks_to_skip % blocks_per_chunk
+    tail_blocks_to_skip = tail_blocks - (blocks_per_chunk - blocks_per_window)
+    return full_windows_to_skip * blocks_per_window + max(0, tail_blocks_to_skip)
+
+
+def transfer_kv_per_object_group(
+    cache_context: GPUCacheContext,
+    block_ids_gpu: list[torch.Tensor],
+    memory_objs: Sequence[MemoryObj | None],
+    object_group_id: int,
+    batch_size: int,
+    skip_first_n_tokens: int,
+    direction: "lmc_ops.TransferDirection",
+) -> None:
+    """Helper function to transfer memory objects of a single object group
+    to/from GPU, with batching support.
+
+    Args:
+        cache_context: The GPU cache context containing the KV cache information.
+        block_ids_gpu: GPU block IDs to retrieve into, indexed by LMCache KV group
+            index. It should satisfy `len(block_ids_gpu[i]) == len(memory_objs) *
+            blocks_per_chunk[i]` for each group `i`.
+            Note that the block IDs list are already on GPU.
+        memory_objs: The list of MemoryObj instances to copy from. It could be
+            None when allocation or retrieval fails. For store (D2H), it should
+            ignore the None entry and continue copying the rest. For retrieve
+            (H2D), it should raise the error and stop copying.
+        object_group_id: Index of the object group being copied.
+        batch_size: The number of memory objects to perform batched copy
+        skip_first_n_tokens: Number of tokens to skip writing at the start of
+            the retrieve range. This avoids overwriting APC-shared GPU blocks that
+            may be read concurrently by other requests.
+        direction: The transfer direction, H2D (retrieve) or D2H (store).
+
+    Raises:
+        ValueError: If it founds None entry in memory_objs when direction is H2D.
+    Note:
+        This function expects the caller to stage the block ids (list[list[int]])
+        into GPU tensors and pass them in as `block_ids_gpu`.
+    """
+    lmcache_chunk_size = cache_context.lmcache_tokens_per_chunk
+    kv_groups_manager = cache_context.kv_layer_groups_manager
+    object_group = kv_groups_manager.object_groups[object_group_id]
+    kernel_group_ids = object_group.kernel_group_indices
+    is_h2d = direction == lmc_ops.TransferDirection.H2D
+
+    sw_size_chunks = kv_groups_manager.get_sw_size_chunks(object_group_id)
+    num_objects_to_skip = 0
+    if sw_size_chunks >= 1 and is_h2d:
+        num_objects_to_skip = max(0, len(memory_objs) - sw_size_chunks)
+        logger.debug(
+            "Detected sliding window for object group %d: "
+            "skipping the first %d objects in the batch",
+            object_group_id,
+            num_objects_to_skip,
+        )
+
+    for start_object_idx, memory_object_batch in batched_iteration_with_skip(
+        memory_objs, batch_size, skip_count=num_objects_to_skip
+    ):
+        if any(mo is None for mo in memory_object_batch):
+            if is_h2d:
+                raise ValueError(
+                    "MemoryObj is None for some objects in the batch, cannot "
+                    "perform H2D copy. memory_object_batch: "
+                    f"{memory_object_batch}"
+                )
+            else:
+                continue
+
+        batch_len = len(memory_object_batch)
+        batch_start_token = start_object_idx * lmcache_chunk_size
+        batch_end_token = batch_start_token + batch_len * lmcache_chunk_size
+
+        effective_start = max(batch_start_token, skip_first_n_tokens)
+        if effective_start >= batch_end_token:
+            continue
+
+        skip_tokens_in_chunk = effective_start - batch_start_token
+
+        # For H2D, copy from CPU to GPU tmp buffers before the kernel launch
+        if is_h2d:
+            for chunk_idx, memory_obj in enumerate(memory_object_batch):
+                lmcache_memcpy_async_h2d(
+                    memory_obj,
+                    cache_context.get_temp_object_group_buffer(
+                        chunk_idx, object_group_id
+                    ),
+                )
+
+        # Do paged KV copy
+        for kernel_group_id in kernel_group_ids:
+            blocks_per_chunk = cache_context.calculate_num_blocks(
+                lmcache_chunk_size, kernel_group_id
+            )
+            tokens_per_window = min(
+                lmcache_chunk_size,
+                kv_groups_manager.get_subchunk_sw_size_tokens(kernel_group_id),
+            )
+            blocks_per_window = cache_context.calculate_num_blocks(
+                tokens_per_window, kernel_group_id
+            )
+
+            # Get the block ids for this chunk
+            start_block_pos = start_object_idx * blocks_per_window
+            end_block_pos = (start_object_idx + batch_len) * blocks_per_window
+
+            block_ids_curr_batch = block_ids_gpu[kernel_group_id][
+                start_block_pos:end_block_pos
+            ]
+
+            # Re-calculate the skip blocks for this kernel group
+            orig_skip_blocks = cache_context.calculate_num_blocks(
+                skip_tokens_in_chunk, kernel_group_id
+            )
+            recalculated_skip_blocks = _recalculate_blocks_to_skip(
+                blocks_per_chunk,
+                blocks_per_window,
+                orig_skip_blocks,
+            )
+
+            # Launch kernel
+            group_kv_pointers = cache_context.get_kernel_group_kv_pointers(
+                kernel_group_id
+            )
+            group_lmcache_chunk_size = cache_context.get_slots_per_chunk_in_sw(
+                kernel_group_id
+            )
+            tmp_gpu_buffers_batched = [
+                cache_context.get_temp_kernel_group_buffer(
+                    i, kernel_group_id
+                ).data_ptr()
+                for i in range(batch_len)
+            ]
+            lmc_ops.multi_layer_block_kv_transfer(
+                group_kv_pointers,
+                tmp_gpu_buffers_batched,
+                block_ids_curr_batch,
+                cache_context.device,
+                direction,
+                cache_context.get_shape_desc(kernel_group_id),
+                group_lmcache_chunk_size,
+                cache_context.engine_kv_format_,
+                recalculated_skip_blocks,
+            )
+
+        # For D2H, copy from GPU tmp buffers to CPU after the kernel launch
+        if not is_h2d:
+            for chunk_idx, memory_obj in enumerate(memory_object_batch):
+                lmcache_memcpy_async_d2h(
+                    cache_context.get_temp_object_group_buffer(
+                        chunk_idx, object_group_id
+                    ),
+                    memory_obj,
+                )
+
+
+@dataclass
+class ContextEntry:
+    """Registered cache context metadata for a single worker instance.
+
+    The concrete type is whatever :func:`create_cache_context` returned
+    for the wrapper list at registration time -- a
+    :class:`GPUCacheContext` for CUDA-IPC wrappers, a
+    :class:`CpuCacheContext` for POSIX-SHM wrappers. Both expose
+    the same ``kv_tensors`` / ``engine_kv_format_`` / ``num_layers`` / ...
+    duck-typed surface, so downstream consumers stay agnostic.
+
+    Args:
+        cache_context: Platform cache context (GPU or CPU) managing
+            shape and pointers to the registered KV cache tensors.
+        model_name: The name of the model associated with this KV cache.
+        world_size: The world size associated with this KV cache.
     """
 
-    metadata: NonGpuContextMetadata
+    cache_context: GPUCacheContext
     model_name: str
     world_size: int
 
@@ -60,9 +406,8 @@ class NonGPUContextEntry:
 class LMCacheDrivenTransferModule:
     """Handles LMCache-driven KV cache transfer operations.
 
-    Owns non-GPU context registrations and provides handlers for
-    register, unregister, prepare/commit store, and prepare/commit retrieve
-    of CPU-serialized KV caches.
+    Owns GPU context registrations and provides handlers for
+    register, unregister, store, and retrieve of GPU KV caches.
 
     Args:
         ctx: The shared engine context.
@@ -70,21 +415,32 @@ class LMCacheDrivenTransferModule:
 
     def __init__(self, ctx: MPCacheServerContext) -> None:
         self._ctx = ctx
-        self._non_gpu_contexts: dict[int, NonGPUContextEntry] = {}
-        self._strategies: dict[int, TransferStrategy] = {}
-        self._pending_shm_writes: dict[
-            tuple[int, IPCCacheServerKey], list[ObjectKey]
-        ] = {}
-        self._pending_shm_reads: dict[
-            tuple[int, IPCCacheServerKey], list[ObjectKey]
-        ] = {}
-        self._pending_shm_lock = threading.Lock()
-        self._shm_pool_info: ShmPoolInfo = self._ctx.shm_pool_info
+        self._cache_contexts: dict[int, ContextEntry] = {}
+
+        # Route finish_write / finish_read_prefetched through a C++ host
+        # callback so the driver thread doesn't acquire the GIL.
+        self._device_host_func_dispatcher = DeviceHostFuncDispatcher()
+        self._device_host_func_dispatcher.register(
+            "finish_write",
+            self._ctx.storage_manager.finish_write,
+            payload_type=list[ObjectKey],
+        )
+        self._device_host_func_dispatcher.register(
+            "finish_read_prefetched",
+            self._ctx.storage_manager.finish_read_prefetched,
+            payload_type=list[ObjectKey],
+        )
+        self._device_host_func_dispatcher.start()
 
     @property
     def context(self) -> MPCacheServerContext:
         """Return the shared engine context. Exposed for testing only."""
         return self._ctx
+
+    @property
+    def cache_contexts(self) -> dict[int, ContextEntry]:
+        """Per-instance GPU context registry."""
+        return self._cache_contexts
 
     def get_handlers(self) -> list[HandlerSpec]:
         """Return handler specs for all request types this module serves.
@@ -95,346 +451,518 @@ class LMCacheDrivenTransferModule:
         """
         return [
             HandlerSpec(
-                RequestType.REGISTER_KV_CACHE_NON_GPU_CONTEXT,
-                self.register_kv_cache_non_gpu_context,
+                RequestType.REGISTER_KV_CACHE,
+                self.register_kv_cache,
                 ThreadPoolType.SYNC,
             ),
             HandlerSpec(
-                RequestType.UNREGISTER_KV_CACHE_NON_GPU_CONTEXT,
+                RequestType.UNREGISTER_KV_CACHE,
                 self.unregister_kv_cache,
                 ThreadPoolType.SYNC,
             ),
             HandlerSpec(
-                RequestType.PREPARE_STORE,
-                self.prepare_store,
+                RequestType.STORE,
+                self.store,
                 ThreadPoolType.AFFINITY,
             ),
             HandlerSpec(
-                RequestType.COMMIT_STORE,
-                self.commit_store,
-                ThreadPoolType.AFFINITY,
-            ),
-            HandlerSpec(
-                RequestType.PREPARE_RETRIEVE,
-                self.prepare_retrieve,
-                ThreadPoolType.AFFINITY,
-            ),
-            HandlerSpec(
-                RequestType.COMMIT_RETRIEVE,
-                self.commit_retrieve,
+                RequestType.RETRIEVE,
+                self.retrieve,
                 ThreadPoolType.AFFINITY,
             ),
         ]
 
     def report_status(self) -> dict:
-        """Return non-GPU transfer module status information.
+        """Return GPU transfer module status information.
 
         Returns:
-            A dict containing registered non-CUDA instance IDs and
-            per-instance context metadata.
+            A dict containing registered GPU instance IDs and
+            per-instance KV cache layout metadata.
         """
-        registered_non_cuda_ids: list[int] = []
-        non_cuda_context_meta: dict[str, dict] = {}
+        registered_gpu_ids: list[int] = []
+        cache_context_meta: dict[str, dict] = {}
 
-        for instance_id, entry in self._non_gpu_contexts.items():
-            registered_non_cuda_ids.append(instance_id)
-            non_cuda_context_meta[str(instance_id)] = {
+        for instance_id, entry in self._cache_contexts.items():
+            registered_gpu_ids.append(instance_id)
+            ctx = entry.cache_context
+            cache_context_meta[str(instance_id)] = {
                 "model_name": entry.model_name,
                 "world_size": entry.world_size,
-                "block_size": entry.metadata.block_size,
-                "use_mla": entry.metadata.use_mla,
+                "kv_cache_layout": ctx.report_status(),
             }
 
         return {
-            "registered_non_cuda_instance_ids": registered_non_cuda_ids,
-            "non_cuda_context_meta": non_cuda_context_meta,
+            "registered_gpu_ids": registered_gpu_ids,
+            "cache_context_meta": cache_context_meta,
         }
 
     def close(self) -> None:
-        """Release resources owned by this module."""
-        self._non_gpu_contexts.clear()
-        self._strategies.clear()
+        """Release GPU resources owned by this module."""
+        # Stop the drain thread before storage_manager.close() so any
+        # in-flight completions reach a live storage manager.
+        self._device_host_func_dispatcher.stop()
 
-    @staticmethod
-    def _make_transfer_key(
-        key: IPCCacheServerKey, instance_id: int
-    ) -> tuple[int, IPCCacheServerKey]:
-        return (instance_id, key)
+        had_contexts = len(self._cache_contexts) > 0
+        for entry in self._cache_contexts.values():
+            entry.cache_context.close()
+        self._cache_contexts.clear()
+        if had_contexts:
+            torch_dev.empty_cache()
 
-    def _resolve_single_group_obj_keys(self, key: IPCCacheServerKey) -> list[ObjectKey]:
-        """Resolve object keys for the single object group used by
-        non-GPU transfers."""
-        return self._ctx.resolve_obj_keys(key, [0])[0]
-
-    def register_kv_cache_non_gpu_context(
+    def register_kv_cache(
         self,
-        payload: RegisterNonGpuContextPayload,
-    ) -> RegisterNonGpuContextResponse:
-        """Register non-CUDA KV layout metadata for non-GPU context mode.
+        instance_id: int,
+        kv_caches: KVCache,
+        model_name: str,
+        world_size: int,
+        engine_type: EngineType,
+        layout_hints: LayoutHints,
+        engine_group_infos: list[EngineGroupInfo],
+    ) -> None:
+        """Register the KV cache tensors for a given GPU instance ID.
 
         Args:
-            payload: Struct containing all registration fields
-                (instance_id, model_name, world_size, block_size,
-                num_layers, hidden_dim_size, dtype_str, use_mla).
-
-        Raises:
-            ValueError: If ``payload.dtype_str`` is not a valid torch dtype name.
+            instance_id: The GPU instance ID (such as PID).
+            kv_caches: The KV cache tensor wrappers from the
+                serving engine.
+            model_name: The name of the model associated with this KV cache.
+            world_size: The world size associated with this KV cache.
+            engine_type: Which serving engine produced the caches.
+                Forwarded to GPUCacheContext for format detection.
+            layout_hints: See LayoutHints.  Forwarded to
+                GPUCacheContext for GPU KV format detection.
+            engine_group_infos: Engine-neutral KV cache group metadata
+                (already msgspec-decoded by the message queue).
         """
-        shm_name = self._shm_pool_info["shm_name"]
-        pool_size = self._shm_pool_info["pool_size"]
-
-        if payload.instance_id in self._non_gpu_contexts:
+        if instance_id in self._cache_contexts:
             logger.warning(
                 "Instance %s's KV cache is already registered, "
                 "skipping the new registration",
-                payload.instance_id,
-            )
-            return RegisterNonGpuContextResponse(shm_name=shm_name, pool_size=pool_size)
-
-        dtype = getattr(torch, payload.dtype_str, None)
-        if dtype is None or not isinstance(dtype, torch.dtype):
-            raise ValueError(
-                f"Invalid dtype_str '{payload.dtype_str}': must be a valid torch dtype "
-                "attribute name (e.g. 'float16' for torch.float16, "
-                "'bfloat16' for torch.bfloat16, 'float32' for torch.float32)."
-            )
-
-        shape = (
-            torch.Size(
-                [payload.num_layers, self._ctx.chunk_size, payload.hidden_dim_size]
-            )
-            if payload.use_mla
-            else torch.Size(
-                [2, payload.num_layers, self._ctx.chunk_size, payload.hidden_dim_size]
-            )
-        )
-        layout_desc = MemoryLayoutDesc(shapes=[shape], dtypes=[dtype])
-        metadata = NonGpuContextMetadata(
-            layout_desc=layout_desc,
-            block_size=payload.block_size,
-            use_mla=payload.use_mla,
-        )
-        self._non_gpu_contexts[payload.instance_id] = NonGPUContextEntry(
-            metadata=metadata,
-            model_name=payload.model_name,
-            world_size=payload.world_size,
-        )
-        strategy: TransferStrategy = create_transfer_strategy(
-            self._ctx.storage_manager,
-            shm_name=shm_name,
-            pool_size=pool_size,
-            pending_writes=self._pending_shm_writes,
-            pending_reads=self._pending_shm_reads,
-            pending_lock=self._pending_shm_lock,
-            transfer_key_factory=self._make_transfer_key,
-        )
-        self._strategies[payload.instance_id] = strategy
-
-        logger.info(
-            "Registered non-GPU context for instance %d (model=%s, world_size=%d)",
-            payload.instance_id,
-            payload.model_name,
-            payload.world_size,
-        )
-
-        self._ctx.layout_desc_registry.register(
-            payload.model_name, payload.world_size, layout_desc
-        )
-        return RegisterNonGpuContextResponse(shm_name=shm_name, pool_size=pool_size)
-
-    def unregister_kv_cache(self, instance_id: int) -> None:
-        """Unregister a non-GPU KV cache context for the given instance ID.
-
-        Args:
-            instance_id: The worker instance identifier.
-        """
-        entry = self._non_gpu_contexts.pop(instance_id, None)
-        if entry is None:
-            logger.warning(
-                "No registered non-GPU context found for instance ID %d",
                 instance_id,
             )
             return
 
-        self._strategies.pop(instance_id, None)
-
-        with self._pending_shm_lock:
-            stale_writes = []
-            stale_reads = []
-            for transfer_key in self._pending_shm_writes:
-                if transfer_key[0] == instance_id:
-                    stale_writes.append(transfer_key)
-            for transfer_key in self._pending_shm_reads:
-                if transfer_key[0] == instance_id:
-                    stale_reads.append(transfer_key)
-
-            write_obj_keys = []
-            for transfer_key in stale_writes:
-                write_obj_keys.append(self._pending_shm_writes.pop(transfer_key))
-
-            read_obj_keys = []
-            for transfer_key in stale_reads:
-                read_obj_keys.append(self._pending_shm_reads.pop(transfer_key))
-
-        for obj_keys in write_obj_keys:
-            if obj_keys:
-                self._ctx.storage_manager.finish_write(obj_keys)
-        for obj_keys in read_obj_keys:
-            if obj_keys:
-                self._ctx.storage_manager.finish_read_prefetched(obj_keys)
-
-        self._ctx.layout_desc_registry.unregister(entry.model_name, entry.world_size)
-        logger.info("Unregistered non-CUDA context for instance ID %d", instance_id)
-
-    @_lmcache_nvtx_annotate
-    def prepare_store(
-        self,
-        key: IPCCacheServerKey,
-        instance_id: int,
-    ) -> PrepareStoreResponse:
-        """Prepare a store operation.
-
-        Args:
-            key: Cache key for the token range to store.
-            instance_id: Worker instance identifier.
-
-        Returns:
-            PrepareStoreResponse with empty slots for pickle mode.
-        """
-        entry = self._non_gpu_contexts.get(instance_id)
-        if entry is None:
-            raise ValueError(
-                f"non-CUDA context not registered for instance ID {instance_id}"
-            )
-        strategy = self._strategies.get(instance_id)
-        if strategy is None:
-            raise ValueError(
-                f"transfer strategy not registered for instance ID {instance_id}"
-            )
-        response = strategy.prepare_store(
-            key=key,
-            instance_id=instance_id,
-            context=entry.metadata,
-            resolve_obj_keys=self._resolve_single_group_obj_keys,
+        cache_context = create_cache_context(
+            kv_caches,
+            self._ctx.chunk_size,
+            layout_hints=layout_hints or None,
+            engine_group_infos=engine_group_infos,
+            engine_type=engine_type,
         )
-        session = self._ctx.session_manager.get_or_create(key.request_id)
-        session.extras["store_start_time"] = time.perf_counter()
-        return response
+        self._cache_contexts[instance_id] = ContextEntry(
+            cache_context=cache_context,
+            model_name=model_name,
+            world_size=world_size,
+        )
+
+        layout_desc = get_layout_desc(
+            cache_context, self._ctx.chunk_size, object_group_id=0
+        )
+        self._ctx.layout_desc_registry.register(model_name, world_size, layout_desc)
+
+        logger.info(
+            "Registered KV cache for GPU ID %d with %d layers",
+            instance_id,
+            cache_context.num_layers,
+        )
+
+    def unregister_kv_cache(self, instance_id: int) -> None:
+        """Unregister the KV cache tensors for a given GPU instance ID.
+
+        Args:
+            instance_id: The GPU instance ID (such as PID).
+        """
+        entry = self._cache_contexts.pop(instance_id, None)
+        if entry is None:
+            logger.warning(
+                "No registered GPU context found for instance ID %d", instance_id
+            )
+            return
+
+        entry.cache_context.close()
+        self._ctx.layout_desc_registry.unregister(entry.model_name, entry.world_size)
+        logger.info("Unregistered KV cache for GPU ID %d", instance_id)
+        torch_dev.empty_cache()
 
     @_lmcache_nvtx_annotate
-    def commit_store(
+    def store(
         self,
         key: IPCCacheServerKey,
         instance_id: int,
-        cpu_data: bytes,
-    ) -> bool:
-        """Commit serialized CPU chunks to storage.
+        gpu_block_ids: list[list[int]],
+        event_ipc_handle: bytes,
+    ) -> tuple[bytes, bool]:
+        """Store the GPU KV cache blocks to CPU.
 
         Args:
-            key: Cache key for the token range to store.
-            instance_id: Worker instance identifier.
-            cpu_data: Pickled list of CPU tensors produced by the worker.
+            key: The IPC key for the KV cache blocks.
+                Must have worker_id != None (worker store operation).
+            instance_id: The GPU instance ID (such as PID).
+            gpu_block_ids: GPU block IDs to store, indexed by LMCache KV
+                group index.
+            event_ipc_handle: The IPC handle of the event to wait on.
 
         Returns:
-            ``True`` when all reserved objects are written, otherwise ``False``.
+            A tuple where the first element is the IPC handle of the event
+            that signals the completion of the store operation, and the second
+            element indicates whether the store operation completed without a
+            fatal error (not whether every requested chunk was stored; see
+            Notes).
 
         Raises:
-            ValueError: If no non-GPU context is registered for the given
-                instance ID.
+            ValueError: If no GPU context is registered for the given instance ID.
+            RuntimeError: If the backend does not support IPC event handles.
+
+        Notes:
+            All-or-nothing. If ``gpu_block_ids`` do not fully cover every chunk
+            ``key`` resolves to for every LMCache group (e.g. a caller/protocol
+            bug), or a copy fails, the whole store is skipped and nothing is
+            committed (logged at WARNING); a subsequent retrieve simply misses
+            and the engine recomputes. The boolean result reports whether the
+            store completed without such a failure.
         """
-        entry = self._non_gpu_contexts.get(instance_id)
+        st = time.perf_counter()
+
+        entry = self._cache_contexts.get(instance_id)
         if entry is None:
-            raise ValueError(
-                f"non-CUDA context not registered for instance ID {instance_id}"
-            )
-        strategy = self._strategies.get(instance_id)
-        if strategy is None:
-            raise ValueError(
-                f"transfer strategy not registered for instance ID {instance_id}"
-            )
-        session = self._ctx.session_manager.get_or_create(key.request_id)
-        st = session.extras.pop("store_start_time", None)
-        result = strategy.commit_store(
-            key=key,
-            instance_id=instance_id,
-            cpu_data=cpu_data,
-            context=entry.metadata,
-            resolve_obj_keys=self._resolve_single_group_obj_keys,
+            raise ValueError(f"No GPU context registered for instance ID {instance_id}")
+        cache_context = entry.cache_context
+        model_name = entry.model_name
+
+        num_object_groups = cache_context.kv_layer_groups_manager.num_object_groups
+        obj_keys_per_obj_group = self._ctx.resolve_obj_keys(
+            key, list(range(num_object_groups))
         )
-        if st is not None and result:
-            num_tokens = (
-                len(self._resolve_single_group_obj_keys(key)) * self._ctx.chunk_size
+        num_chunks = len(obj_keys_per_obj_group[0])
+
+        # NOTE: different engine groups may have different block sizes, so
+        # ``blocks_per_chunk[i]`` is the number of blocks in one chunk for
+        # group ``i``.
+        blocks_per_chunk = [
+            cache_context.calculate_num_blocks(self._ctx.chunk_size, group_idx)
+            for group_idx in range(
+                cache_context.kv_layer_groups_manager.num_kernel_groups
             )
+        ]
+
+        with (
+            torch_dev.device(cache_context.device),
+            torch_dev.stream(cache_context.stream),
+        ):
+            check_interprocess_event_support()
+            event = torch_dev.Event(interprocess=True)
+
+            # Fail closed: every LMCache group must have block IDs covering all
+            # chunks. A short list (e.g. a caller/protocol bug) would otherwise
+            # drive the transfer kernel to read out-of-bounds GPU memory, so skip
+            # the whole store and commit nothing rather than caching a partial or
+            # garbage entry. A later request can store it once the block IDs are
+            # complete. Checked on the raw block ids, before cutting drops the
+            # per-chunk blocks that sliding-window groups do not need.
+            if any(
+                len(group_block_ids) < num_chunks * bpc
+                for group_block_ids, bpc in zip(
+                    gpu_block_ids, blocks_per_chunk, strict=True
+                )
+            ):
+                logger.warning(
+                    "STORE block ID underflow for request_id=%s: each group needs "
+                    "num_chunks * blocks_per_chunk block IDs for %d chunks "
+                    "(per-group blocks_per_chunk=%s); skipping the store.",
+                    key.request_id,
+                    num_chunks,
+                    blocks_per_chunk,
+                )
+                event.record()
+                return event.ipc_handle(), False
+
+            block_ids_per_group_gpu = downsample_and_stage_block_ids(
+                cache_context, gpu_block_ids
+            )
+
+            if not hasattr(torch_dev.Event, "from_ipc_handle"):
+                raise RuntimeError(
+                    f"Backend '{torch_device_type}' does not support IPC event "
+                    "handles (Event.from_ipc_handle not available). "
+                    "Multiprocess IPC requires CUDA."
+                )
+            vllm_event = torch_dev.Event.from_ipc_handle(
+                cache_context.device, event_ipc_handle
+            )
+            vllm_event.wait(stream=cache_context.stream)
+
+            # CPU-synchronous sentinel: a GPU store is about to be enqueued.
+            # Must be published via publish() (not publish_on_stream) so the
+            # drain thread sees it before MP_REQUEST_END can race MP_STORE_END.
+            self._ctx.event_bus.publish(
+                Event(
+                    event_type=EventType.MP_STORE_SUBMITTED,
+                    session_id=key.request_id,
+                    metadata={"device": str(cache_context.device)},
+                )
+            )
+
+            self._ctx.event_bus.publish_on_stream(
+                cache_context.cupy_stream,
+                Event(
+                    event_type=EventType.MP_STORE_START,
+                    session_id=key.request_id,
+                    metadata={
+                        "device": str(cache_context.device),
+                        "engine_id": instance_id,
+                        "model_name": model_name,
+                    },
+                ),
+            )
+
+            reserved_dict: dict[ObjectKey, MemoryObj] = {}
+            all_dict: dict[ObjectKey, MemoryObj] = {}
+            total_bytes: int = 0
+            store_succeeded = False
+            try:
+                for obj_group_id in range(num_object_groups):
+                    obj_keys = obj_keys_per_obj_group[obj_group_id]
+                    layout_desc = get_layout_desc(
+                        cache_context,
+                        self._ctx.chunk_size,
+                        object_group_id=obj_group_id,
+                    )
+                    reserved_dict = self._ctx.storage_manager.reserve_write(
+                        obj_keys, layout_desc, "new"
+                    )
+                    all_dict.update(reserved_dict)
+                    if reserved_dict:
+                        total_bytes += next(
+                            iter(reserved_dict.values())
+                        ).get_size() * len(reserved_dict)
+
+                    # Keys not in reserved_dict (skipped by the storage manager)
+                    # become None entries; the helper skips them for D2H.
+                    memory_objs: list[MemoryObj | None] = [
+                        reserved_dict.get(obj_key) for obj_key in obj_keys
+                    ]
+
+                    # NOTE: batch_size must stay 1 for store.
+                    transfer_kv_per_object_group(
+                        cache_context,
+                        block_ids_per_group_gpu,
+                        memory_objs,
+                        object_group_id=obj_group_id,
+                        batch_size=1,
+                        skip_first_n_tokens=0,
+                        direction=lmc_ops.TransferDirection.D2H,
+                    )
+
+                store_succeeded = True
+            except Exception:
+                logger.exception("Cannot store keys due to exception")
+                return event.ipc_handle(), False
+            finally:
+                event.record()
+                # Fail closed: commit the reserved objects only when every chunk
+                # copied successfully; otherwise the whole store is skipped.
+                stored_count = len(all_dict) if store_succeeded else 0
+                if stored_count:
+                    submit_callback_to_stream(
+                        cache_context.cupy_stream,
+                        "finish_write",
+                        list(all_dict.keys()),
+                    )
+                else:
+                    total_bytes = 0
+                self._ctx.event_bus.publish_on_stream(
+                    cache_context.cupy_stream,
+                    Event(
+                        event_type=EventType.MP_STORE_END,
+                        session_id=key.request_id,
+                        metadata={
+                            "stored_count": stored_count,
+                            "device": str(cache_context.device),
+                            "engine_id": instance_id,
+                            "model_name": model_name,
+                            "total_bytes": total_bytes,
+                        },
+                    ),
+                )
+
+        ed = time.perf_counter()
+        if stored_count:
             logger.info(
                 "Stored %d tokens in %.3f seconds",
-                num_tokens,
-                time.perf_counter() - st,
+                num_chunks * self._ctx.chunk_size,
+                ed - st,
             )
-        return result
+        return event.ipc_handle(), True
 
     @_lmcache_nvtx_annotate
-    def prepare_retrieve(
+    def retrieve(
         self,
         key: IPCCacheServerKey,
         instance_id: int,
-    ) -> PrepareRetrieveResponse:
-        """Retrieve prefetched chunks and return serialized CPU tensors.
+        gpu_block_ids: list[list[int]],
+        event_ipc_handle: bytes,
+        skip_first_n_tokens: int = 0,
+    ) -> tuple[bytes, bool]:
+        """Retrieve the CPU KV cache and put into GPU blocks.
 
         Args:
-            key: Cache key for the token range to retrieve.
-            instance_id: Worker instance identifier.
+            key: The IPC key for the KV cache blocks.
+                Must have worker_id != None (worker retrieve operation).
+            instance_id: The GPU instance ID (such as PID).
+            gpu_block_ids: GPU block IDs to retrieve into, indexed by LMCache
+                KV group index.
+            event_ipc_handle: The IPC handle of the event to wait on.
+            skip_first_n_tokens: Number of tokens to skip writing at
+                the start of the retrieve range. This avoids overwriting
+                APC-shared GPU blocks that may be read concurrently by other
+                requests.
 
         Returns:
-            PrepareRetrieveResponse with serialized data on hit.
+            A tuple where the first element is the IPC handle of the event
+            that signals the completion of the retrieve operation, and the
+            second element indicates whether the key was successfully retrieved.
 
         Raises:
-            ValueError: If no non-GPU context is registered for the given
-                instance ID.
+            ValueError: If no GPU context is registered for the given instance ID.
         """
-        strategy = self._strategies.get(instance_id)
-        if strategy is None:
-            raise ValueError(
-                f"transfer strategy not registered for instance ID {instance_id}"
-            )
-        response = strategy.prepare_retrieve(
-            key=key,
-            instance_id=instance_id,
-            resolve_obj_keys=self._resolve_single_group_obj_keys,
+        st = time.perf_counter()
+
+        entry = self._cache_contexts.get(instance_id)
+        if entry is None:
+            raise ValueError(f"No GPU context registered for instance ID {instance_id}")
+        cache_context = entry.cache_context
+        model_name = entry.model_name
+
+        num_object_groups = cache_context.kv_layer_groups_manager.num_object_groups
+        obj_keys_per_obj_group = self._ctx.resolve_obj_keys(
+            key, list(range(num_object_groups))
         )
-        session = self._ctx.session_manager.get_or_create(key.request_id)
-        session.extras["retrieve_start_time"] = time.perf_counter()
-        return response
+        num_chunks = len(obj_keys_per_obj_group[0])
 
-    @_lmcache_nvtx_annotate
-    def commit_retrieve(
-        self,
-        key: IPCCacheServerKey,
-        instance_id: int,
-    ) -> bool:
-        """Finalize a retrieve operation.
+        # CPU-synchronous sentinel: a GPU retrieve is about to be enqueued.
+        # Must be published via publish() (not publish_on_stream) so the
+        # drain thread sees it before MP_REQUEST_END can race MP_RETRIEVE_END.
+        self._ctx.event_bus.publish(
+            Event(
+                event_type=EventType.MP_RETRIEVE_SUBMITTED,
+                session_id=key.request_id,
+                metadata={"device": str(cache_context.device)},
+            )
+        )
 
-        Args:
-            key: Cache key (unused for pickle).
-            instance_id: Worker instance identifier (unused for pickle).
+        self._ctx.event_bus.publish_on_stream(
+            cache_context.cupy_stream,
+            Event(
+                event_type=EventType.MP_RETRIEVE_START,
+                session_id=key.request_id,
+                metadata={
+                    "device": str(cache_context.device),
+                    "engine_id": instance_id,
+                    "model_name": model_name,
+                },
+            ),
+        )
 
-        Returns:
-            Always ``True``.
-        """
-        strategy = self._strategies.get(instance_id)
-        if strategy is None:
-            raise ValueError(
-                f"transfer strategy not registered for instance ID {instance_id}"
+        blocks_per_chunk = [
+            cache_context.calculate_num_blocks(self._ctx.chunk_size, group_idx)
+            for group_idx in range(
+                cache_context.kv_layer_groups_manager.num_kernel_groups
             )
-        session = self._ctx.session_manager.get_or_create(key.request_id)
-        st = session.extras.pop("retrieve_start_time", None)
-        result = strategy.commit_retrieve(key=key, instance_id=instance_id)
-        if st is not None:
-            num_tokens = (
-                len(self._resolve_single_group_obj_keys(key)) * self._ctx.chunk_size
+        ]
+
+        with (
+            torch_dev.device(cache_context.device),
+            torch_dev.stream(cache_context.stream),
+        ):
+            check_interprocess_event_support()
+            event = torch_dev.Event(interprocess=True)
+
+            # Fail closed: a short block-id list would drive the transfer
+            # kernel to write out-of-bounds GPU memory. Checked on the raw
+            # block ids, before cutting drops the per-chunk blocks that
+            # sliding-window groups do not need.
+            if any(
+                len(group_block_ids) < num_chunks * bpc
+                for group_block_ids, bpc in zip(
+                    gpu_block_ids, blocks_per_chunk, strict=True
+                )
+            ):
+                logger.error(
+                    "RETRIEVE block ID underflow for request_id=%s: each group "
+                    "needs num_chunks * blocks_per_chunk block IDs for %d "
+                    "chunks (per-group blocks_per_chunk=%s); skipping the "
+                    "retrieve.",
+                    key.request_id,
+                    num_chunks,
+                    blocks_per_chunk,
+                )
+                event.record()
+                return event.ipc_handle(), False
+
+            # Cut and stage all block_ids to GPU once before the transfer
+            block_ids_per_group_gpu = downsample_and_stage_block_ids(
+                cache_context, gpu_block_ids
             )
-            logger.info(
-                "Retrieved %d tokens in %.3f seconds",
-                num_tokens,
-                time.perf_counter() - st,
-            )
-        return result
+
+            prefetched_keys: list[ObjectKey] = []
+            total_bytes = 0
+            try:
+                for obj_group_id in range(num_object_groups):
+                    obj_keys = obj_keys_per_obj_group[obj_group_id]
+                    with self._ctx.storage_manager.read_prefetched_results(
+                        obj_keys
+                    ) as memory_objs:
+                        if not memory_objs or len(memory_objs) != len(obj_keys):
+                            logger.error("Some keys not found during retrieve!")
+                            return event.ipc_handle(), False
+
+                        total_bytes += sum(mo.get_size() for mo in memory_objs)
+
+                        transfer_kv_per_object_group(
+                            cache_context,
+                            block_ids_per_group_gpu,
+                            memory_objs,
+                            object_group_id=obj_group_id,
+                            batch_size=cache_context.max_batch_size,
+                            skip_first_n_tokens=skip_first_n_tokens,
+                            direction=lmc_ops.TransferDirection.H2D,
+                        )
+                        # Extend only after the copy is enqueued: on exception,
+                        # read_prefetched_results releases this group's locks
+                        # itself, and a key must not be released twice.
+                        prefetched_keys.extend(obj_keys)
+            except Exception:
+                logger.exception("Cannot retrieve keys due to exception")
+                return event.ipc_handle(), False
+            finally:
+                event.record()
+                if prefetched_keys:
+                    submit_callback_to_stream(
+                        cache_context.cupy_stream,
+                        "finish_read_prefetched",
+                        prefetched_keys,
+                    )
+                self._ctx.event_bus.publish_on_stream(
+                    cache_context.cupy_stream,
+                    Event(
+                        event_type=EventType.MP_RETRIEVE_END,
+                        session_id=key.request_id,
+                        metadata={
+                            "retrieved_count": len(prefetched_keys),
+                            "device": str(cache_context.device),
+                            "engine_id": instance_id,
+                            "model_name": model_name,
+                            "cache_salt": key.cache_salt,
+                            "total_bytes": total_bytes,
+                        },
+                    ),
+                )
+        tokens_retrieved = num_chunks * self._ctx.chunk_size
+        ed = time.perf_counter()
+        logger.info(
+            "Retrieved %d tokens in %.3f seconds",
+            tokens_retrieved,
+            ed - st,
+        )
+
+        return event.ipc_handle(), True
