+++ lmcache/v1/multiprocess/server.py
@@ -119,7 +119,7 @@ class MPCacheServer:
     def gpu_contexts(self) -> dict[int, GPUCacheContext] | None:
         """Used by ``/kvcache/check``; unwraps :class:`ContextEntry`."""
         for module in self._modules:
-            if isinstance(module, EngineDrivenTransferModule):
+            if isinstance(module, LMCacheDrivenTransferModule):
                 return {i: e.cache_context for i, e in module.cache_contexts.items()}
         return None
 
@@ -174,12 +174,12 @@ def _build_modules(
     ]
 
     if mp_config.supported_transfer_mode == "gpu":
-        modules.append(EngineDrivenTransferModule(ctx))
+        modules.append(LMCacheDrivenTransferModule(ctx))
     elif mp_config.supported_transfer_mode == "non_gpu":
-        modules.append(LMCacheDrivenTransferModule(ctx))
-    elif mp_config.supported_transfer_mode == "auto":
         modules.append(EngineDrivenTransferModule(ctx))
+    elif mp_config.supported_transfer_mode == "auto":
         modules.append(LMCacheDrivenTransferModule(ctx))
+        modules.append(EngineDrivenTransferModule(ctx))
     else:
         raise ValueError(
             f"Unsupported supported_transfer_mode '{mp_config.supported_transfer_mode}'"
@@ -212,7 +212,7 @@ def _build_modules(
         from lmcache.v1.multiprocess.modules.blend_v3 import BlendV3Module
 
         gpu_transfer = next(
-            m for m in modules if isinstance(m, EngineDrivenTransferModule)
+            m for m in modules if isinstance(m, LMCacheDrivenTransferModule)
         )
         lookup_module = next(m for m in modules if isinstance(m, LookupModule))
         # Opt-in: enabled only when LMCACHE_COORDINATOR_URL is set; otherwise
