+++ tests/v1/multiprocess/test_batched_iteration_with_skip.py
@@ -5,7 +5,7 @@
 import pytest
 
 # First Party
-from lmcache.v1.multiprocess.modules.engine_driven_transfer import (
+from lmcache.v1.multiprocess.modules.lmcache_driven_transfer import (
     batched_iteration_with_skip,
 )
